[C-5766] Profile pagination
